third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/block-container-block-end-last-child-with-border-expected.txt

This is a testharness.js-based test.
[FAIL] .collapsed 2
  assert_equals: \n<item data-offset-y="59" class="collapsed" style="margin-block: 5px 8px;"></item>\noffsetTop expected 59 but got 79
Harness: the test ran to completion.
